fix(web): group insight results by platform
